notebooks/requirements.txt
pyttsx3==2.90
gtts==2.4.0
pygame==2.5.2
requests==2.31.0
urllib3==2.0.7
